work88/work89 audit: build fixes, B-1/B-2/B-3 UB fixes, rule-of-five, CAS null fixes
Key fixes:
- B-1: compareExchangeAtomic nullptr type mismatch → static_cast<DSPCore*>(nullptr) (5 locations)
- B-2: SnapshotCoordinator ISRAuthorityClass.h include, RetireEnqueueResult → bool return
- B-3: updateFade() lambda capture fix [this, useDryAsOld]
- PsychoacousticDither move ctor/assign deleted (rule of five)
- myGeneration dead variable removed, recoverDSPPtr removed

Docs:
- REPAIR_PLAN.md: refined B-1/B-2/B-3 + U-3/U-4
- CONSOLIDATED_BUG_LIST.md + IMPLEMENT_CHECKLIST.md
- mini_bugs_checked/: BUG-011..046 consolidated from work88 root
- work89/: new bug reports BUG-047..065

+++ src/TruePeakDetector.cpp
@@ -99,8 +99,8 @@ double TruePeakDetector::processBlock(const double* dataL, const double* dataR, 
         return 0.0;
 
     double* work = upsampleBuffer.get();
-    const int up1Samples = numSamples * 2;
-    const int up2Samples = numSamples * 4;
+    const size_t up1Samples = static_cast<size_t>(numSamples) * 2;
+    const size_t up2Samples = static_cast<size_t>(numSamples) * 4;
 
     // オフセット: work 領域のレイアウト
     //   [ Stage0 L | Stage0 R | Stage1 L | Stage1 R ]
